Update cache when a file no longer has displayable todos
When the last displayable todo item is completed or otherwise removed
from a file, the item is not removed from the plugin's UI. This is
because the logic for updating the `itemsByFile` cache only considers
files with displayable todos.

To resolve this issue, we need to consider all updated files when
updating the `itemsByFile` cache, regardless of whether they have
displayable todos. We can do this by grouping todos by file when
searching for todos in the `parseTodos` method, then simply updating
`itemsByFile` with each map entry returned by `parseTodos`, even if the
todos list is empty.

GitHub issue #112